backend/reporting_service/requirements.txt
fastapi==0.95.2        # A stable version that works well with Pydantic v1
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
pydantic==1.10.12      # Crucial: Downgrade Pydantic to v1 for compatibility with FastAPI 0.95.2
pymongo==4.3.3         # Specific version for pymongo
motor==3.1.2 
aio-pika==9.5.5